requirements.txt
Django
djangorestframework
mysqlclient
click
